language: consolidate language-specific coalescers
All language ecosystem coalescers shared the same logic: skip layers
without repos, associate packages with repos. The gobin coalescer had
unnecessary repo name/PackageDB prefix filtering since scanners already
ensure only relevant packages appear.
