+++ language/coalescer_test.go
@@ -1,4 +1,4 @@
-package nodejs
+package language
 
 import (
 	"strconv"
@@ -12,9 +12,14 @@ import (
 func TestCoalescer(t *testing.T) {
 	t.Parallel()
 	ctx := test.Logging(t)
+	repoName := "npm"
 	coalescer := &coalescer{}
 	pkgs := test.GenUniquePackages(6)
-	repo := []*claircore.Repository{&Repository}
+	repo := []*claircore.Repository{{
+		ID:   "1",
+		Name: repoName,
+		URI:  "https://www.npmjs.com/",
+	}}
 	layerArtifacts := []*indexer.LayerArtifacts{
 		{
 			Hash: test.RandomSHA256Digest(t),
@@ -48,11 +53,10 @@ func TestCoalescer(t *testing.T) {
 		t.Fatalf("received error from coalesce method: %v", err)
 	}
 	// Expect 0-5 to have gotten associated with the repository.
-	for i := range pkgs {
+	for i := range pkgs[:5] {
 		es, ok := ir.Environments[strconv.Itoa(i)]
-		if !ok && i == 5 {
-			// Left out the last package.
-			continue
+		if !ok {
+			t.Errorf("missing environment for package %d", i)
 		}
 		e := es[0]
 		if len(e.RepositoryIDs) == 0 {
@@ -60,9 +64,12 @@ func TestCoalescer(t *testing.T) {
 		}
 		for _, id := range e.RepositoryIDs {
 			r := ir.Repositories[id]
-			if got, want := r.Name, Repository.Name; got != want {
+			if got, want := r.Name, repoName; got != want {
 				t.Errorf("got: %q, want: %q", got, want)
 			}
 		}
+	}
+	if _, ok := ir.Environments[strconv.Itoa(5)]; ok {
+		t.Error("expected last package to be excluded (no repo in its layer)")
 	}
 }
